Allow to set span name manually in datadog
In datadog a span name is supposed to be its categore and a resource
name is displayed next to the span name. In order to make our spans look
good we can't really set a custom name like for Honeycomb, but I would
also like to control it somehow, so it's consistent with Honeycomb. This
commit makes the collector use the type field if it exists for setting
the span name